Initial commit: Publishing old SerialWifi plugin
Also comes with an GCodeLibrary, which makes it possible to do 'dry' prints for test cases.

In addition there is also a 'PowerManagement', which is caring about suspends.
BUT: There is already a PR with it at Ultimaker/Cura, which needs to be rewritten into QDBus, also it probably needs to be renamed into SystemManagement, since it can be extended in future with other system related stuff.